♻️ 🚨 Fix `react-hooks/exhaustive-deps` rule of ESLint with `eslint-interactive`
Several `useEffect` hooks in the `src/` folder were missing dependencies, which has now been corrected.
In the `Feed` component, the `handleScroll` function was moved inside the `useEffect` to make it clearer that this function needs to be recreated whenever the `isLoading` or `hasMorePosts` variables change.
From 63 problems, there are now 58.