♻️ 🚨 Fix `no-unused-vars` rule of ESLint
There are no longer unused variables in the `src/` folder, lightening and harmonizing the code.
Previously, in the `Post` component, the `expand` variable of the `ExpandMore` sub-component was extracted from the props passed in the function arguments, even though the `Button` component doesn't interpret it. Consequently, the whole destructuring has been safely removed.
From 2 problems, there are now 0.