🔧 Add `import/no-extraneous-dependencies` rule to ESLint configuration
This rule is not included by default in StandardJS, but it helps enforce maintainability in the codebase. It detects imports of packages that are not listed as direct dependencies of the app in `package.json`. In fact, sub dependencies can be imported just like any other, but with the risk of breaking the app if their parent dependencies are updated or removed.
With this rule, ESLint now detects 25 new problems.