➕ 🚨 Add `prop-types` package as Yarn dependency, fixing violations of the `import/no-extraneous-dependencies` rule in ESLint
`PropTypes` has been used in the app since the start of the project, but it was never installed as a direct dependency. Instead, it was only a sub-dependency of `react-router-hash-link` and MUI.
In fact, sub-dependencies can be imported just like any other, but this comes with the risk of breaking the app if their parent dependencies are updated or removed.
From 25 problems, there are now 0.