♻️ Refactor exports to always use inline named syntax wherever possible
Both end-of-file and inline export styles were used in the app, so the codebase needed to be harmonized. Unfortunately, there is no ESLint rule to enforce either of these two practices, requiring to a manual refactor.
Inline exports are the most commonly used form, especially with components. They prevent repeating the component name, and eliminate the need to jump back and forth between its declaration and export.