🔧 Add `react/no-unused-prop-types` rule to ESLint configuration
This rule is not included by default in `eslint-plugin-react`, but it helps enforce a clean codebase. It detects `PropTypes` definitions that don't match any of the component props, which is good to spot misspelled ones, but more importantly, it helps remove unnecessary ones.
With this rule, ESLint now detects 1 new problem.