⚰️ Remove `src/data/` folder
It contained mocked data that was used during early stages of the app's development, before the API was ready.
However, it's now unnecessary because there is no swap mechanism to swap the mocked data for the API. Since the API must be installed on every development machine, it can never be unreachable, making such a system redundant.
This `data/` folder was prone to confusion, so it's better to delete it.