🏗️ Move `src/components/Theme/index.jsx` file to `src/styles/muiONetworkTheme.js`
This file does not contain a React component; it is purely a configuration file that creates a MUI theme. For these reasons, it has been moved to the `src/styles` folder (so outside the `components` folder), renamed to a clearer camelCase name and its extension has been changed from `.jsx` to `.js`.