♻️ Refactor `ActivityFeed` and its sub-components to use `ProfileUserIdContext` instead of prop drilling
Previously, the user ID for user's profile page was extracted from the URL in each component that required it, with some components also applying additional processing (a `parseInt()` with or without a radix). This approach broke the "single source of truth" principle, so now this ID - along with any necessary processing - is provided by a dedicated context, improving maintainability and developer experience.