♻️ Rename `userId` and `userIdUrl` variables for better clarity and consistency in user profile-related components
Specifically, `userIdUrl` was too contextual: it's more important to clearly indicate that this ID belongs to the currently displayed user profile than to specify where the data was extracted from.
However, in the case of the `fetchPosts` thunk, we are outside the profile page context and the user ID is passed as a function parameter, so naming it `userId` is sufficient here.