♻️ Refactor activity-feed selectors to use `createSelector`
It's the recommended way to create selectors in Redux Toolkit, as it uses memoization for improved performance.
Using such selectors is necessary for the next major change in the feature-based architecture: splitting the `feed` slice into each `activity-feed` sub-feature. Data will be normalized, and `createSelector` is designed to work very efficiently with this kind of data structure.