⚡️ Refactor `ActivityFeed` to reduce re-rendering
Previously, the posts loop in `ActivityFeed` used full posts objects. This caused unnecessary re-rendering when adding or removing comments or reactions, as such interaction mutates the `commentIds` and `reactionIds` properties of the affected post. For example, adding a reaction triggers a re-render of the reactions list, which in turn was causing a re-render of the post, ultimately re-rendering the entire posts list.
Now, a separate array of the posts IDs is used, provided by Redux's `createEntityAdapter`, which avoids the issue.